% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspCP-package.R
\docType{package}
\name{dspCP-package}
\alias{dspCP}
\alias{dspCP-package}
\title{dspCP: Dynamic Shrinkage Process and Change Points Detection}
\description{
Implements methods for modeling and detecting change points in time series data using a Bayesian state-space approach with dynamic shrinkage. The model utilizes a autoregressive stochastic volatility process within a time-varying parameter model to estimate smoothly varying trends with large, isolated interruptions. Change points are estimated using a threshold parameter. The model can deal with heteroscedastic measurement error, and jointly identifies both local outliers and change points.
}
\author{
\strong{Maintainer}: Haoxuan Wu \email{hw399@cornell.edu}

Authors:
\itemize{
  \item Daniel R. Kowal \email{daniel.kowal@rice.edu}
  \item David S. Matteson \email{matteson@cornell.edu}
}

}
\keyword{internal}
