% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{identify_cp}
\alias{identify_cp}
\title{Identify changepoints from the output of ABCO}
\usage{
identify_cp(D, mcmc_output, cp_thres = 0.5)
}
\arguments{
\item{D}{the degree of differencing for changepoint, \code{D=1}}

\item{mcmc_output}{MCMC outputs from ABCO, needs "omega" and "r"}

\item{cp_thres}{Percentage of posterior samples needed to declare a changepoint}
}
\description{
Identify changepoints from the output of ABCO
}
