% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{make.signal}
\alias{make.signal}
\title{Generate univariate signals of different type}
\usage{
make.signal(name, n = 1024, snr = Inf)
}
\arguments{
\item{name}{name of the signal to be generated}

\item{n}{length of the series}

\item{snr}{desired signal-to-noise ratio}
}
\description{
Using code from the archived \code{wmtsa} package
}
