% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sample_mat_c}
\alias{sample_mat_c}
\title{Wrapper function for C++ call for sample mat, check pre-conditions to prevent crash}
\usage{
sample_mat_c(row_ind, col_ind, mat_val, mat_l, num_inp, linht, rd, D)
}
\arguments{
\item{row_ind}{list of the row indices to fill in the bandsparse matrix}

\item{col_ind}{list of the columns indices to fill in the bandsparse matrix}

\item{mat_val}{list of the values to fill in the bandsparse matrix}

\item{mat_l}{dimension of the band-sparse matrix}

\item{num_inp}{number of non-zero elements in the bandsparse matrix}

\item{linht}{\code{T-D} vector of linear term in the sampler}

\item{rd}{\code{T-D} vector of standard normal noise samples}

\item{D}{the degree of differencing for changepoint}
}
\description{
Wrapper function for C++ call for sample mat, check pre-conditions to prevent crash
}
