% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abco.R
\name{t_create_loc}
\alias{t_create_loc}
\title{Initializer for location indices for filling in band-sparse matrix}
\usage{
t_create_loc(len, D)
}
\arguments{
\item{len}{length of the diagonal of the band-sparse matrix}

\item{D}{number of super-diagonals to include for the band-sparse}
}
\value{
a list containing
\itemize{
\item the row indices \code{r} and
\item the column indices \code{c}
}
}
\description{
Create row and column indices for locations of symmetric band-sparse matrix.
Starts with the locations of the diagonal, proceed with upper-diagonals,
followed by lower-diagonals.
}
