% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abco.R
\name{t_initEvolParams_no}
\alias{t_initEvolParams_no}
\title{Initialize the evolution error variance parameters}
\usage{
t_initEvolParams_no(y, D, omega)
}
\arguments{
\item{y}{the \code{T} vector of time series observations}

\item{D}{degree of differencing (D = 1, or D = 2)}

\item{omega}{\code{T} vector of evolution errors}
}
\value{
List of relevant components: \code{sigma_wt}, the \code{T} vector of evolution standard deviations,
and additional parameters associated with the DHS priors.
}
\description{
Compute initial values for evolution error variance parameters under the
dynamic horseshoe prior
}
