% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abco.R
\name{t_initEvolZeta_ps}
\alias{t_initEvolZeta_ps}
\title{Initialize the anomaly component parameters}
\usage{
t_initEvolZeta_ps(zeta)
}
\arguments{
\item{zeta}{\code{T} vector of initial estimates.}
}
\value{
List of relevant components: \code{sigma_wt}, the \code{T} vector of standard deviations,
and additional parameters for inverse gamma priors (shape and scale).
}
\description{
Compute initial values for either a horseshoe prior or horseshoe+ prior
for the anomaly component.
}
