% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abco.R
\name{t_initSV}
\alias{t_initSV}
\title{Initialize the stochastic volatility parameters}
\usage{
t_initSV(omega)
}
\arguments{
\item{omega}{\code{T} vector of errors}
}
\value{
List of relevant components: \code{sigma_wt}, the \code{T} vector of standard deviations,
and additional parameters (unconditional mean, AR(1) coefficient, and standard deviation).
}
\description{
Compute initial values for normal stochastic volatility parameters.
The model assumes an AR(1) for the log-volatility.
}
