% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abco.R
\name{t_sampleEvolZeta_ps}
\alias{t_sampleEvolZeta_ps}
\title{Sampler for the anomaly component parameters}
\usage{
t_sampleEvolZeta_ps(omega, evolParams)
}
\arguments{
\item{omega}{\code{T} vector of errors}

\item{evolParams}{list of parameters to be updated}
}
\value{
List of relevant components in \code{evolParams}:  \code{sigma_wt},
the \code{T} vector of standard deviations, and additional parameters for inverse gamma priors (shape and scale).
}
\description{
Compute one draw of the anomaly component parameters.
}
